+++ src/names.h
@@ -65,5 +65,6 @@ tree_t query_name(nametab_t *tab, ident_t name);
 tree_t find_std(nametab_t *tab);
 
 type_t solve_types(nametab_t *tab, tree_t expr, type_t constraint);
+type_t solve_condition(nametab_t *tab, tree_t expr, type_t constraint);
 
 #endif // _NAMES_H